Implement status and ticket action handling in Application class
- Added `_handle_status_action` and `_handle_ticket_action` methods to the Application class to manage status-related and ticket button actions across different integrations.
- Updated Mattermost, Slack, and Telegram application classes to utilize the new action handling methods, improving code reuse and consistency.
- Enhanced logging for button actions to provide clearer insights into user interactions.
- Introduced a utility function in tests to standardize button configurations for Telegram, reducing code duplication in test cases.
