+++ app/im/slack/slack_application.py
@@ -115,20 +115,6 @@ class SlackApplication(Application):
             self._track_async_task(asyncio.create_task(self.post_unassignment_notification(incident_)))
             incident_.release()
 
-    async def _handle_status_action(self, incident_):
-        """Handle status-related button actions"""
-        if incident_.status_enabled:
-            logger.info(f'Incident {incident_.uuid} -> button STATUS pressed (disabled)')
-            incident_.status_enabled = False
-        else:
-            logger.info(f'Incident {incident_.uuid} -> button STATUS pressed (enabled)')
-            incident_.status_enabled = True
-
-    async def _handle_ticket_action(self, incident_, queue_):
-        """Handle Ticket button action"""
-        logger.info(f'Incident {incident_.uuid} -> button TICKET pressed')
-        self._track_async_task(asyncio.create_task(self.handle_ticket_button(incident_, queue_)))
-
     async def buttons_handler(self, payload, incidents, queue_, route):
         config = get_config()
         if payload.get('token') != config.slack_verification_token:
@@ -148,7 +134,7 @@ class SlackApplication(Application):
             if action['name'] == 'chain':
                 await self._handle_chain_action(incident_, user_id, queue_, incidents)
             elif action['name'] == 'status':
-                await self._handle_status_action(incident_)
+                await self._handle_status_action(incident_, not incident_.status_enabled)
             elif action['name'] == 'ticket':
                 await self._handle_ticket_action(incident_, queue_)
