+++ tests/test_im/test_telegram/test_telegram_application.py
@@ -229,14 +229,13 @@ class TestTelegramApplication:
 
     def test_create_thread_payload(self, app_config, channels, users):
         """Test _create_thread_payload method."""
+        from tests.utils import create_telegram_buttons_mock
+        
         app = self.create_telegram_app(app_config, channels, users)
 
         with patch('app.im.telegram.telegram_application.buttons') as mock_buttons:
-            mock_buttons.__getitem__.side_effect = lambda x: {
-                'chain': {'takeit': {'text': 'Take It', 'callback_data': 'start_chain'}},
-                'status': {'enabled': {'text': 'Status', 'callback_data': 'start_status'}},
-                'ticket': {'create': {'text': '📌', 'callback_data': 'ticket'}}
-            }[x]
+            buttons_config = create_telegram_buttons_mock()
+            mock_buttons.__getitem__.side_effect = lambda x: buttons_config[x]
 
             result = app._create_thread_payload(-1001234567890, "body", "header", "5312241539987020022", "firing")
 
@@ -272,20 +271,13 @@ class TestTelegramApplication:
 
     def test_update_thread_payload(self, app_config, channels, users):
         """Test update_thread_payload method."""
+        from tests.utils import create_telegram_buttons_mock
+        
         app = self.create_telegram_app(app_config, channels, users)
 
         with patch('app.im.telegram.telegram_application.buttons') as mock_buttons:
-            mock_buttons.__getitem__.side_effect = lambda x: {
-                'chain': {
-                    'takeit': {'text': 'Take It', 'callback_data': 'start_chain'},
-                    'release': {'text': 'Release', 'callback_data': 'stop_chain'}
-                },
-                'status': {
-                    'enabled': {'text': 'Status', 'callback_data': 'start_status'},
-                    'disabled': {'text': 'Status', 'callback_data': 'stop_status'}
-                },
-                'ticket': {'create': {'text': '📌', 'callback_data': 'ticket'}}
-            }[x]
+            buttons_config = create_telegram_buttons_mock()
+            mock_buttons.__getitem__.side_effect = lambda x: buttons_config[x]
 
             result = app.update_thread_payload(-1001234567890, "123456/789012", "body", "header", "5312241539987020022",
                                               "firing", True, True)
@@ -309,20 +301,13 @@ class TestTelegramApplication:
 
     def test_update_thread_payload_with_task_link(self, app_config, channels, users):
         """Test update_thread_payload method when task link exists - button should be removed."""
+        from tests.utils import create_telegram_buttons_mock
+        
         app = self.create_telegram_app(app_config, channels, users)
 
         with patch('app.im.telegram.telegram_application.buttons') as mock_buttons:
-            mock_buttons.__getitem__.side_effect = lambda x: {
-                'chain': {
-                    'takeit': {'text': 'Take It', 'callback_data': 'start_chain'},
-                    'release': {'text': 'Release', 'callback_data': 'stop_chain'}
-                },
-                'status': {
-                    'enabled': {'text': 'Status', 'callback_data': 'start_status'},
-                    'disabled': {'text': 'Status', 'callback_data': 'stop_status'}
-                },
-                'ticket': {'create': {'text': '📌', 'callback_data': 'ticket'}}
-            }[x]
+            buttons_config = create_telegram_buttons_mock()
+            mock_buttons.__getitem__.side_effect = lambda x: buttons_config[x]
 
             # Pass task_link to verify button is NOT included
             result = app.update_thread_payload(-1001234567890, "123456/789012", "body", "header", "5312241539987020022",
